Move Routing::Redirect and Routing::Uri into routing.rb
